# Optimize the bytes encoding at node